Discovered that you need a bit extra on a x64 OS
My previous work with this was on a 32 bit os Fedora 22 instance. Because of
that I never discovered that I needed to set argtypes and that I needed to
handle pointers a bit more delicately

It is also possible this is coming from changes in python.ctypes or numpy,
but what I did should be backwards compatible. It is also possible (though
it seems unlikely) that this is because of the change in the GCC ABI but I
thought that python should be agnostic to that

Signed-off-by: James Till Matta <jamesmatta@gmail.com>